BISERVER-8503 - Moving getRolesForUser and getUserForRole in the list service resource
